+++ src/theme/colors.ts
@@ -1,0 +1,41 @@
+import { Palette } from "./palette.enum";
+
+/** Colors for the VSCode UI */
+export default {
+  colors: {
+    "activityBar.activeBorder": Palette.Blue_turquoise,
+    "activityBar.background": Palette.Dire_wolf,
+    "activityBar.foreground": Palette.Blue_turquoise,
+    "activityBar.inactiveForeground": Palette.Inactive_grey, // plugin icons on the left bar
+    "activityBarBadge.background": Palette.Pink_bright,
+    "activityBarBadge.foreground": Palette.White_regular,
+    "button.background": Palette.Violet_dark,
+    "editor.background": Palette.Dire_wolf,
+    "editor.foreground": Palette.White_regular,
+    "editorGroupHeader.tabsBackground": Palette.Dire_wolf,
+    "editorGutter.background": Palette.Dire_wolf,
+    "editorIndentGuide.activeBackground": Palette.Blue_turquoise,
+    "editorIndentGuide.background": Palette.Jet,
+    "editorInlayHint.parameterForeground": Palette.Grey_bright,
+    "editorLineNumber.activeForeground": Palette.Violet_dark,
+    "editorLineNumber.foreground": Palette.Goshawk_grey,
+    "editorOverviewRuler.border": Palette.Jet,
+    "editorRuler.foreground": Palette.Goshawk_grey,
+    "list.activeSelectionBackground": Palette.Dimmed_blue,
+    "menu.selectionBackground": Palette.Blue_turquoise,
+    "scrollbar.shadow": Palette.Dire_wolf,
+    "sideBar.background": Palette.Dire_wolf,
+    "sideBar.border": Palette.Jet,
+    "sideBarSectionHeader.background": Palette.Dire_wolf,
+    "sideBarSectionHeader.border": Palette.Jet,
+    "sideBarTitle.foreground": Palette.White_regular,
+    "statusBar.background": Palette.Violet_dark,
+    "tab.activeBackground": Palette.Dimmed_pink,
+    "tab.activeBorder": Palette.Pink_bright,
+    "tab.activeBorderTop": Palette.Pink_bright,
+    "tab.inactiveBackground": Palette.Dire_wolf,
+    "titleBar.activeBackground": Palette.Dire_wolf,
+    "tree.indentGuidesStroke": Palette.Blue_turquoise,
+    focusBorder: Palette.Blue_turquoise,
+  },
+};