+++ src/theme/token_colors_variants/cpp-colors.ts
@@ -1,0 +1,11 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "C++ string",
+    scope: ["string.quoted.double.cpp", "string.quoted.single.cpp"],
+    settings: {
+      foreground: Palette.Yello_light,
+    },
+  },
+];