+++ src/theme/token_colors_variants/go-colors.ts
@@ -1,0 +1,11 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "GO string",
+    scope: ["string.quoted.double.go", "string.quoted.single.go"],
+    settings: {
+      foreground: Palette.Yello_light,
+    },
+  },
+];