+++ src/theme/token_colors_variants/ruby-colors.ts
@@ -1,0 +1,11 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "RUBY string",
+    scope: ["string.quoted.double.ruby", "string.quoted.single.ruby"],
+    settings: {
+      foreground: Palette.Yello_light,
+    },
+  },
+];