+++ src/theme/token_colors_variants/sass-colors.ts
@@ -1,0 +1,11 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "SCSS variable",
+    scope: ["variable.scss"],
+    settings: {
+      foreground: Palette.Blue_turquoise,
+    },
+  },
+];