+++ src/theme/token_colors_variants/typescript-colors.ts
@@ -1,0 +1,15 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "TS string",
+    scope: [
+      "string.quoted.double.ts",
+      "string.quoted.single.ts",
+      "string.template.ts",
+    ],
+    settings: {
+      foreground: Palette.Yello_light,
+    },
+  },
+];