+++ src/theme/token_colors_variants/yaml-colors.ts
@@ -1,0 +1,11 @@
+import { Palette } from "../palette.enum";
+
+export default [
+  {
+    name: "YAML string",
+    scope: ["string.quoted.double.yaml", "string.quoted.single.yaml"],
+    settings: {
+      foreground: Palette.Yello_light,
+    },
+  },
+];