+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ---
 
+## [1.0.0] - 15-10-2022
+
+### Changed
+- Migrated project from pure Json to Typescript.
+- Dimmed editor indent guides
+- Darkened background
+- Lightened softtened pink and green colors
+---
+
 ## [0.7.1] - 15-10-2022
 
 ### Fixed